# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agent api documentation",
        "version": "1.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth user"
                ],
                "summary": "Auth user",
                "description": "Returns user data and token",
                "operationId": "Registration user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegistrationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth user"
                ],
                "summary": "Auth user",
                "description": "Returns user data and token",
                "operationId": "Login user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth user"
                ],
                "summary": "Logout user",
                "description": "Logout user",
                "operationId": "Logout user",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get list of clients",
                "description": "Returns list of clients",
                "operationId": "a29537dc61149cb33bd0ef259ffc4248",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients/{client}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get client information",
                "description": "Return client data",
                "operationId": "80aefec375e76113904bed48494ad1d3",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "Client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notification": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get notification list",
                "description": "Return notifications data",
                "operationId": "01822d9e4711cd19d8287379642f2d09",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "query",
                        "description": "Client id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Create notification",
                "description": "Returns user data",
                "operationId": "eb7b9488e098e6f9facd2c7188b2aa54",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendNotificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notification/{notification}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get notification information",
                "description": "Return notification data",
                "operationId": "4a3c35420269a36342fe8f9956298323",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "description": "Notification id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserLoginRequest": {
                "title": "User login",
                "description": "User login",
                "properties": {
                    "email": {
                        "title": "User email",
                        "description": "User email",
                        "type": "string",
                        "example": "john.doe@mail.com"
                    },
                    "password": {
                        "title": "User password",
                        "description": "User password",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User login"
                }
            },
            "UserRegistrationRequest": {
                "title": "User registration",
                "description": "User registration",
                "properties": {
                    "name": {
                        "title": "User name",
                        "description": "User name",
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "title": "User email",
                        "description": "User email",
                        "type": "string",
                        "example": "john.doe@mail.com"
                    },
                    "password": {
                        "title": "User password",
                        "description": "User password",
                        "type": "string",
                        "example": "12345678"
                    },
                    "password_confirmation": {
                        "title": "User password confirmation",
                        "description": "User password confirmation",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User registration"
                }
            },
            "SendNotificationRequest": {
                "title": "Send notification request",
                "description": "Send notification request body data",
                "required": [
                    "notification"
                ],
                "properties": {
                    "notification": {
                        "title": "notification",
                        "description": "Notification list",
                        "type": "array",
                        "items": {
                            "properties": {
                                "clientId": {
                                    "description": "Id of client",
                                    "type": "integer",
                                    "example": 1
                                },
                                "channel": {
                                    "description": "Channel type (Sms or Email)",
                                    "type": "string",
                                    "example": "email"
                                },
                                "content": {
                                    "description": "Content of notification",
                                    "type": "string",
                                    "example": "Content of notification"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}